gcp_functions/translation/functions/requirements.txt
blinker==1.6.3
CacheControl==0.13.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudevents==1.9.0
cryptography==41.0.4
deprecation==2.1.0
firebase-admin==6.2.0
firebase-functions==0.1.1
Flask==2.3.3
Flask-Cors==4.0.0
functions-framework==3.4.0
google-api-core==1.34.0
google-api-python-client==2.103.0
google-auth==1.35.0
google-auth-httplib2==0.1.1
google-cloud-core==1.7.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-cloud-translate==2.0.1
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.48.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
watchdog==3.0.0
Werkzeug==3.0.0
